migrated aggregation code to multinet.ts for use with the clickable buttons for rendering
